+++ src/App.tsx
@@ -2,7 +2,7 @@ import { BrowserRouter as Router } from "react-router-dom";
 
 import { RootLayout } from "@/app/ui/layout.tsx";
 
-import PostsManagerPage from "./pages/PostsManagerPage.tsx";
+import PostsManagerPage from "./pages/posts-manager-page.tsx";
 
 const App = () => {
   return (
