--- src/widgets/post-table/index.ts
@@ -1,1 +1,0 @@
-export { PostsTableWidget } from "./ui/post-table";
